US Country as Actor/Recipient
Contains daily and monthly analysis when the US as an actor and the US as a recipient.  Analysis is divided into two sections within the file - US Actor and US Recipient. Various country-based and action-based analyses are done. Daily (& Monthly) average sentiment of each action/country when US is an actor/recipient.