US Government as Actor/Recipient
Contains daily and monthly analysis when the US government is the actor or recipient with U.S. business. The whole analysis is divided into two sections within the file - US Government is Actor and US Government is Recipient. When US government is recipient, US Business will be the target actor (and vice-versa).  Daily (& Monthly) average sentiment of each action when US government is an actor/recipient.